+++ neuronquant/__init__.py
@@ -2,6 +2,11 @@
 QuanTrade
 '''
 
+__author__    = 'Xavier Bruhiere'
+__copyright__ = 'Xavier Bruhiere'
+__licence__   = 'Apache 2.0'
+__version__   = '0.4'
+
 import ai
 import gears
 import data
@@ -11,7 +16,7 @@ import utils
 __all__ = [
         'ai',
         'gears',
-        'data', 
+        'data',
         'graph',
         'utils'
-        ]
+]