+++ neuronquant/gears/algorithms.py
@@ -41,11 +41,11 @@ class BuyAndHold(TradingAlgorithm):
         signals = dict()
 
         ''' ----------------------------------------------------------    Scan   --'''
+        import ipdb; ipdb.set_trace()
         if self.frame_count == 4:
             for ticker in data:
                 signals[ticker] = data[ticker].price
-        import ipdb; ipdb.set_trace()
-        if self.frame_count == 5:
+        if self.frame_count == 6:
             self.stop_trading()
 
         ''' ----------------------------------------------------------   Orders  --'''
@@ -59,14 +59,14 @@ class BuyAndHold(TradingAlgorithm):
     #TODO Anyway should find a more elegant way
     def stop_trading(self):
         ''' Convenient method to stop calling user algorithm and just finish the simulation'''
-        self.logger('Trader out of the market')
+        self.logger.info('Trader out of the market')
         #NOTE Selling every open positions ?
         # Saving the portfolio in database, eventually for reuse
         self.manager.save_portfolio(self.portfolio)
 
         # Closing generator
         self.date_sorted.close()
-        self.set_datetime(self.sim_params.last_close)
+        #self.set_datetime(self.sim_params.last_close)
         self.done = True
 
 
@@ -96,6 +96,8 @@ class DualMovingAverage(TradingAlgorithm):
 
         self.short_mavgs = []
         self.long_mavgs  = []
+
+        self.init = True
 
     def handle_data(self, data):
         ''' ----------------------------------------------------------    Init   --'''