+++ neuronquant/tmpdata/__init__.py
@@ -1,0 +1,3 @@
+from datafeed import DataFeed
+
+__all__ = ['DataAgent']