+++ neuronquant/utils/__init__.py
@@ -6,4 +6,4 @@ __VERSION__ = 0.1
 
 from logger import LogSubsystem, log, setup, remote_setup, color_setup
 from dates import epochToDate
-from utils import to_dict
+from utils import to_dict, apply_mapping