+++ scripts/installation/install_r_packages.R
@@ -5,7 +5,9 @@
 #However should detect if user did not already setup it, and if not
 install.packages('quantmod', repos='http://cran.univ-paris1.fr/')
 install.packages('RMySQL', repos='http://cran.univ-paris1.fr/')
+install.packages('RSQLite', repos='http://cran.univ-paris1.fr/')
 install.packages('PerformanceAnalytics', repos='http://cran.univ-paris1.fr/')
 install.packages('shiny', repos='http://cran.univ-paris1.fr/')
 install.packages('rzmq', repos='http://cran.univ-paris1.fr/')
 install.packages('futile.logger', repos='http://cran.univ-paris1.fr/')
+install.packages('optparse', repos='http://cran.univ-paris1.fr/')